requirements.txt
fyers-apiv3
pandas
numpy
ta
apscheduler
pyyaml
requests
yfinance
